rename elements already in components dir
