+++ components/code-example/element.js
@@ -3,6 +3,7 @@ import { LitElement, html } from "lit";
 import { createRef, ref } from "lit/directives/ref.js";
 import { unsafeHTML } from "lit/directives/unsafe-html.js";
 
+import "../copy-button/element.js";
 import styles from "./element.css?lit";
 
 export class MDNCodeExample extends LitElement {
