+++ pages/search/index.js
@@ -1,7 +1,6 @@
 import { html } from "lit";
 
 import { PageLayout } from "../../components/page-layout/index.js";
-import "../../components/site-search/index.js";
 
 /**
  * @param {Fred.Context} context
