+++ types/element-map.d.ts
@@ -4,7 +4,11 @@ declare global {
   interface HTMLElementTagNameMap {
     "mdn-button": import("../components/button/element.js").MDNButton;
     "mdn-code-example": import("../components/code-example/element.js").MDNCodeExample;
+    "mdn-color-theme": import("../components/color-theme/element.js").MDNColorTheme;
+    "mdn-copy-button": import("../components/copy-button/element.js").MDNCopyButton;
+    "mdn-content-feedback": import("../components/content-feedback/element.js").MDNContentFeedback;
     "mdn-search-modal": import("../components/search-modal/element.js").MDNSearchModal;
+    "mdn-site-search": import("../components/site-search/element.js").MDNSiteSearch;
   }
 }
